Add mechanism to replace native 'backup' method
The native backup functionality in Neovim (and Vim) does not work
consistently with the handling of swapfiles. For swapfiles, a trailing
`//` in the value of 'directory' will cause swapfiles to be saved with
a name derived from the whole path to the original file, with `/`
characters replaced by `%`. This prevents files with the same name in
different directories from overwriting each others' swapfile.
Unfortunately, the 'backupdir' option has no such provision, making it
useless if you want to avoid conflivting backup files when using `set
backup`

The function `lib#backupdir()` provides a replacement for the
traditional 'backupdir' options that supports the `//` convention.